+++ CHANGES.md
@@ -1,3 +1,6 @@
+#### Version: 1.9 (built with Ghidra 10.0.4)
+- Ghidra Debugger can be used to debug games running in PPSSPP
+
 #### Version: 1.8 (built with Ghidra 10.0.4)
 - Updated to Ghidra 10.0.4
 - Added initial support for decompiling VFPU instructions